Update testing-components-with-graphql.md
I could just not fully understand what is going on here, and I thought this would be the best place to make this note since it would just be a small update to the docs.

But, StaticQuery doesn't seem to actually mock the data, it will throw a "cannot read properties of undefined" when a component one layer down is trying to access this mocked data. 

So if I'm testing HomePage.js that  looks something like this

```
<div>
    <MyComponent />
</div>
```

and MyComponent has some static query that wants to access, say, the metadata or something in allmarkdownremark or something you want to mock in the static query before the test runs. It won't be able to access that data if you use StaticQuery, but seems to work fine with the useStaticQuery hook.

Again, I know I don't fully understand the intricate difference between the two kinds of ways to static query, but I know that the one that is in the current docs right now doesn't seem to work with the explained example above.

If I'm wrong or misunderstand how this all weeks, an explanation would be great! I'm still trying to learn my way through all of gatsby and graphQL.